% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInstrumentName.R
\name{GetInstrumentName}
\alias{GetInstrumentName}
\title{Get Instrument Name}
\usage{
GetInstrumentName(x)
}
\arguments{
\item{x}{a \code{.raw} file}
}
\value{
a \code{list} of four elements
}
\description{
Extract instrumment identification fields
}
