% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrecursorMasses.R
\name{GetPrecursorMasses}
\alias{GetPrecursorMasses}
\title{Get Precursor Masses}
\usage{
GetPrecursorMasses(x)
}
\arguments{
\item{x}{a \code{.raw} file}
}
\value{
a numeric vector of MS Level 2 precursor mass values
}
\description{
Get Precursor Masses
}
