% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleInfo.R
\name{GetSampleInfo}
\alias{GetSampleInfo}
\title{Get Sample Information}
\usage{
GetSampleInfo(x)
}
\value{

}
\description{
Extract all fields contained in the \code{SampleInformation} block of Thermo \code{.raw} files. \code{.raw} files
are accessed using the cross-platform C# RawFileReader.
}
