% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetScanFilters.R
\name{GetScanFilters}
\alias{GetScanFilters}
\title{Get Scan Filters}
\usage{
GetScanFilters(x)
}
\arguments{
\item{x}{a \code{.raw} file}
}
\value{
a \code{tibble} of scan filters
}
\description{
Extract all unique scan filters available in the \code{.raw} file
}
